% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{baseFit}
\alias{baseFit}
\title{The function to get the fit of W}
\usage{
baseFit(meth, unmeth, p, q, e, size)
}
\arguments{
\item{meth}{detection read counts data}

\item{unmeth}{total read counts minus detection read counts data}

\item{p}{estimated p vector}

\item{q}{estimated q vector}

\item{e}{estimated e vector}

\item{size}{size factor}
}
\value{
a list of the fit variable of meth data and unmeth data
}
\description{
The function to get the fit of W
}
