% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateW}
\alias{calculateW}
\title{The function to get estimated W}
\usage{
calculateW(data, size, e)
}
\arguments{
\item{data}{the data of read count}

\item{size}{size factor}

\item{e}{e factor}
}
\value{
a estimated W
}
\description{
The function to get estimated W
}
