% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateZ}
\alias{calculateZ}
\title{The function to get estimated Z}
\usage{
calculateZ(q, p, size, e)
}
\arguments{
\item{q}{estimated abundance vector}

\item{p}{the estimated p}

\item{size}{the estimated size factor}

\item{e}{the estimated e vector}
}
\value{
estimated Z
}
\description{
The function to get estimated Z
}
