% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateE}
\alias{estimateE}
\title{estimate E function.}
\usage{
estimateE(meth, unmeth, size, q)
}
\arguments{
\item{meth}{detection read counts data}

\item{unmeth}{total read counts minus detection read counts data}

\item{size}{size factor}

\item{q}{estimate the abundance of feature base on all data/ false treatment data}
}
\value{
an estimated e vector
}
\description{
estimate E function.
}
