% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateP}
\alias{estimateP}
\title{estimate P function.}
\usage{
estimateP(meth, unmeth, size)
}
\arguments{
\item{meth}{detection read counts data}

\item{unmeth}{total read counts minus detection read counts data}

\item{size}{size factor}
}
\value{
an estimated P vector
}
\description{
estimate P function.
}
