% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateQ}
\alias{estimateQ}
\title{estimate Q function.}
\usage{
estimateQ(total_control, total_test, s_control, s_test)
}
\arguments{
\item{total_control}{total read counts for control data}

\item{total_test}{total read counts for test data}

\item{s_control}{size factor for control data}

\item{s_test}{size factor for test data}
}
\value{
an estimated q vector
}
\description{
estimate Q function.
}
