% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fittedW}
\alias{fittedW}
\title{The function to get final estimated Z}
\usage{
fittedW(p, q, fit)
}
\arguments{
\item{p}{estimated p level}

\item{q}{estimated q abundance}

\item{fit}{the fit list in the locfitw function}
}
\value{
the final estimated W
}
\description{
The function to get final estimated Z
}
