% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getBaseEstimate}
\alias{getBaseEstimate}
\title{The function to calculate all estimate variable for replicate happen on test data.}
\usage{
getBaseEstimate(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  s_control,
  s_test
)
}
\arguments{
\item{meth_control}{control IP data}

\item{meth_test}{treated IP data}

\item{unmeth_control}{control Input data}

\item{unmeth_test}{treated Input data}

\item{s_control}{size factor for control data}

\item{s_test}{size factor for test data}
}
\value{
a list
}
\description{
The function to calculate all estimate variable for replicate happen on test data.
}
