% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEstimate}
\alias{getEstimate}
\title{The function to calculate all estimate variable.}
\usage{
getEstimate(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  s_control,
  s_test
)
}
\arguments{
\item{meth_control}{detect read counts from false treatment}

\item{meth_test}{detect read counts from correct treatment}

\item{unmeth_control}{total read counts minus detection read counts data from false treatment}

\item{unmeth_test}{total read counts minus detection read counts data from correct treatment}

\item{s_control}{size factor for control data}

\item{s_test}{size factor for test data}
}
\value{
a list
}
\description{
The function to calculate all estimate variable.
}
