% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locfitW}
\alias{locfitW}
\title{The function to local fit the W}
\usage{
locfitW(p, q, w)
}
\arguments{
\item{p}{estimated p level}

\item{q}{estimated q abundance}

\item{w}{estimated W}
}
\value{
a locfit list
}
\description{
The function to local fit the W
}
