% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotDispersion}
\alias{plotDispersion}
\title{plot the dispersion fit}
\usage{
plotDispersion(fit_meth, fit_unmeth, path)
}
\arguments{
\item{fit_meth}{fit list of the meth from the locfitW function or
the baseFit fucntion}

\item{fit_unmeth}{fit list of the unmeth from the locfitW function or
the baseFit fucntion}

\item{path}{the path}
}
\value{
a pdf with dispersion fit plot
}
\description{
plot the dispersion fit
}
