% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quadNBtest}
\alias{quadNBtest}
\title{calculate the p value}
\usage{
quadNBtest(t1, t, n2, mu2_t, mu2_c, size2_t, size2_c)
}
\arguments{
\item{t1}{is the total number of detected treated data}

\item{t}{is the total number of detected data}

\item{n2}{is the total number of treated data}

\item{mu2_t}{mean of t_2}

\item{mu2_c}{mean of c_2}

\item{size2_t}{size factor of t_2}

\item{size2_c}{size factor of c_2}
}
\value{
p-value
}
\description{
calculate the p value
}
