% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification_level.R
\name{quantification_level}
\alias{quantification_level}
\title{The function to get the three quantification levels}
\usage{
quantification_level(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  size.factor = NA,
  q = NA,
  adjust_1 = 1e-05,
  adjust_2 = 0.01
)
}
\arguments{
\item{meth_control}{detect methylation read counts from false treatment
(row is gene and column is single cell)}

\item{meth_test}{detect methyation read counts from correct treatment
(row is gene and column is single cell)}

\item{unmeth_control}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{unmeth_test}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{size.factor}{NA or a data.frame with one column is control size factor and test size factor}

\item{q}{expression aboundance}

\item{adjust_1}{adjust parameter for RR}

\item{adjust_2}{adjust parameter for OR}
}
\value{
a list c("methylation level treated","methylation level control","log2 Risk Ratio",
"log2 Odds Ratio","TCR")

1.The first part of the list is the meth proportion data frame (row is gene
and column is single cell) in the test cells.
2.The second part of the list is the mean meth proportion vector in the control cells.
3.The third part of the list is the log2 risk ratio data frame for the test cells.
4.The forth part of the list is the log2 odds ratio data frame for the test cells.
5.The fifth part of the list is the TCR for the test cells.

For NA in the return list:
\enumerate{
\item NA in the treated methylation level, it is caused by that the test cells have no reads
in both methylation reads and un methylation reads
\item NA in the control methylation level, it is caused by that the all control cells have
no reads in both methylation reads and un methylation reads
\item NA in the risk ratio, the methylation proportion is NA on the test cell or control cells group.
\item NA in the TCR, is similar to the risk ratio.
}
}
\description{
The function to get the three quantification levels
}
\examples{
 data <- simulateData(n_Sites=100,test_num = 2,control_num = 10)
 res <- quantification_level(data[[1]],data[[2]],data[[3]],data[[4]])
}
