% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SIGMR_rep_test}
\alias{SIGMR_rep_test}
\title{This is the sub function for SIGMRA_similarity_test function}
\usage{
SIGMR_rep_test(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  size.factor = NA,
  plot.dispersion = FALSE,
  output.dir = NA,
  remove.false = TRUE
)
}
\arguments{
\item{meth_control}{detect read counts from false treatment}

\item{meth_test}{detect read counts from correct treatment}

\item{unmeth_control}{total read counts minus detection read counts data from false treatment}

\item{unmeth_test}{total read counts minus detection read counts data from correct treatment}

\item{size.factor}{TRUE or FALSE}

\item{plot.dispersion}{whether plot or not}

\item{output.dir}{the path}

\item{remove.false}{remove some false positive or not}
}
\value{
a list c("detect proportion treated","detect proportion control","log2 Risk Ratio",
"log2 Odds Ratio","p value","aboundance","adjusted p value")
}
\description{
This is the sub function for SIGMRA_similarity_test function
}
