% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGMR_similarity_test.R
\name{SIGMR_similarity_test}
\alias{SIGMR_similarity_test}
\title{This function serves as the central component for single-cell RNA methylation
site analysis by selecting the control group based on the similarity of each
test cell. Users can choose the number of control cells, which are selected
based on their similarity to the test cells. While it is possible to choose
the number of test cells, it is not recommended, and the default is set to 1.}
\usage{
SIGMR_similarity_test(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  expression = NA,
  method = "Pearson",
  num_test = 1,
  num_control,
  size.factor = NA,
  plot.dispersion = FALSE,
  output.dir = NA,
  remove.false = TRUE
)
}
\arguments{
\item{meth_control}{detect methylation read counts from false treatment
(row is gene and column is single cell)}

\item{meth_test}{detect methyation read counts from correct treatment
(row is gene and column is single cell)}

\item{unmeth_control}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{unmeth_test}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{expression}{the expression matrix for compare the similarity, if this
variable is miss, we use read counts}

\item{method}{calculate similarity method c("Euclidean distance","Manhattan distance",
"Maximum distance","Pearson","Spearman")}

\item{num_test}{number of cells in the test group (usual be 1 unless you have
replicates of the test cells)}

\item{num_control}{number of cells in the control group}

\item{size.factor}{TRUE or FALSE}

\item{plot.dispersion}{whether plot or not}

\item{output.dir}{the path}

\item{remove.false}{remove some false positive or not}
}
\value{
a list  c("detect proportion treated","detect proportion control","log2 Risk Ratio",
"log2 Odds Ratio","p value","abundance","adjusted p value")

1.The first part of the list is the meth proportion data frame (row is gene
and column is single cell) in the test cells.
2.The second part of the list is the mean meth proportion data frame (row is gene
and column is single cell) in the control cells.
3.The third part of the list is the log2 risk ratio data frame for the test cells.
4.The forth part of the list is the log2 odds ratio data frame for the test cells.
5.The fifth part of the list is the p value data frame for the test cells.
6.The sixth part of the list is the estimated gene abundance data frame.
7.The seventh part of the list is the adjusted p value data frame for the test cells.
For NA in the return list:
\enumerate{
\item NA in the treated detect proportion, it is caused by that the test cells have no reads
in both methylation reads and un methylation reads
\item NA in the control detect proportion, it is caused by that the all control cells have
no reads in both methylation reads and un methylation reads
\item NA in the risk ratio, the methylation proportion is NA on the test cell or control cells group,
or the methylation proportion is 0 in both test cell and control group.
\item NA in the odds ratio, is similar to the risk ratio
\item NA in the p value, it is only caused by the that control detect proportion is NA and the
treated detect proportion is not NA. If the treated detect proportion is NA the p-value is 1.
}
}
\description{
This function serves as the central component for single-cell RNA methylation
site analysis by selecting the control group based on the similarity of each
test cell. Users can choose the number of control cells, which are selected
based on their similarity to the test cells. While it is possible to choose
the number of test cells, it is not recommended, and the default is set to 1.
}
\examples{
  data <- simulateData (test_num=1,control_num=30)

  meth_control=data[[1]];meth_test=data[[2]];unmeth_control=data[[3]];unmeth_test=data[[4]]

  # use Pearson
  res1 <- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Pearson", num_test=1,num_control=3,size.factor=NA,plot.dispersion=FALSE,
  output.dir = NA,remove.false = TRUE)

  # use Euclidean distance
  res2 <- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Euclidean distance",num_test=1,num_control=3,size.factor=NA,
  plot.dispersion=FALSE,   output.dir = NA,remove.false = TRUE)

 # use expression and Euclidean distance
 expression <- matrix (runif(31000,0,1),nrow=1000,ncol=31)
 res3<- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
 expression=expression,method="Euclidean distance",num_test=1,num_control=3,
 size.factor=NA,plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)



  data <- simulateData (test_num=7,control_num=30)

  meth_control=data[[1]];meth_test=data[[2]];unmeth_control=data[[3]];unmeth_test=data[[4]]

  # use Pearson
  res1 <- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Pearson",num_test=1,num_control=3,size.factor=NA,plot.dispersion=FALSE,
  output.dir = NA,remove.false = TRUE)

  # use Euclidean distance
  res2 <- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Euclidean distance",num_test=1,num_control=3,size.factor=NA,
  plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)

  res3 <- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Pearson", num_test=3,num_control=3,size.factor=NA,plot.dispersion=
  FALSE,output.dir = NA,remove.false = TRUE)

  # use Euclidean distance with 3 replicates of test data
  res4 <- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Euclidean distance",num_test=3,num_control=3,size.factor=NA,
  plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)


 # use expression and Euclidean distance
 expression <- matrix (runif(37000,0,1),nrow=1000,ncol=37)
 res5<- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
 expression=expression,method="Euclidean distance",num_test=1,num_control=3,
 size.factor=NA,plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)

 # use expression and Euclidean distance with 3 replicates of test data
 res6<- SIGMR_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
 expression=expression,method="Euclidean distance",num_test=3,num_control=3,
 size.factor=NA,plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)
}
