% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRM_cluster_test.R
\name{SigRM_cluster_test}
\alias{SigRM_cluster_test}
\title{This is the main function if you know the cluster of the cells}
\usage{
SigRM_cluster_test(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  expression = NA,
  cluster,
  similarity = FALSE,
  method = "Pearson",
  num_control = 3,
  size.factor = NA,
  plot.dispersion = FALSE,
  output.dir = NA,
  remove.false = TRUE,
  withq = TRUE,
  method_dispersion = "locfit",
  adjust_1 = 1e-05,
  adjust_2 = 0.01,
  one_side = TRUE
)
}
\arguments{
\item{meth_control}{detect methylation read counts from false treatment
(row is gene and column is single cell)}

\item{meth_test}{detect methyation read counts from correct treatment
(row is gene and column is single cell)}

\item{unmeth_control}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{unmeth_test}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{expression}{the expression matrix for similarity, if this variable is miss,
we use read counts}

\item{cluster}{cluster index factor vector, so you need cluster the cell before
put the data into this function}

\item{similarity}{TRUE or FALSE, if the cluster data do not have control data,
we use all control data or use similarity to choose some cells.}

\item{method}{calculate similarity method}

\item{num_control}{number of cells in the control group (work for similarity)}

\item{size.factor}{TRUE or FALSE (have size factor or not)}

\item{plot.dispersion}{whether plot or not}

\item{output.dir}{the path}

\item{remove.false}{remove some false positive or not}

\item{withq}{with the expression abundance or not}

\item{method_dispersion}{the method to estimate the dispersion ("unbiased_estimate","mle","locfit")}

\item{adjust_1}{adjust parameter for RR}

\item{adjust_2}{adjust parameter for OR}

\item{one_side}{p value is one side or both sides}
}
\value{
a list c("methylation level treated","methylation level control","log2 Risk Ratio",
"log2 Odds Ratio","p value","aboundance","adjusted p value","TCR")

1.The first part of the list is the meth proportion data frame (row is gene
and column is single cell) in the test cells.
2.The second part of the list is the mean meth proportion data frame (row is gene
and column is single cell) in the control cells.
3.The third part of the list is the log2 risk ratio data frame for the test cells.
4.The forth part of the list is the log2 odds ratio data frame for the test cells.
5.The fifth part of the list is the p value data frame for the test cells.
6.The sixth part of the list is the estimated gene abundance data frame.
7.The seventh part of the list is the adjusted p value data frame for the test cells.
8.The eighth part of the list is the TCR for the test cells.

For NA in the return list:
\enumerate{
\item NA in the treated methylation level, it is caused by that the test cells have no reads
in both methylation reads and un methylation reads
\item NA in the control methylation level, it is caused by that the all control cells have
no reads in both methylation reads and un methylation reads
\item NA in the risk ratio, the methylation proportion is NA on the test cell or control cells group.
\item NA in the TCR, is similar to the risk ratio.
\item NA in the p value, it is only caused by the that control methylation level is NA and the
treated methylation level is not NA.
}
}
\description{
This function serves as the central component for single-cell RNA methylation
site analysis by clustering cells based on gene expression or read counts,
encompassing both test and control groups. Before calling this function, it
is essential to cluster the data. Ideally, each cluster comprises a mix of test
and control cells. Within these clusters, control cell data serves as a
reference, enabling the assessment of test cell behavior. This approach relies
on control cells with similar gene expression patterns as a baseline for
comparison. The method systematically evaluates the behavior of test cells
within their clusters. However, if a cluster lacks control cells, you can
either choose several of the most similar control cells for each test cell
in the cluster (by calling the SigRM_similarity_test()) or use all control
cells as the background information.
}
\examples{
  set.seed(1)
  data <- simulateData (test_num=1,control_num=30)

  meth_control=data[[1]];meth_test=data[[2]];unmeth_control=data[[3]];unmeth_test=data[[4]]

  # simulate the cluster vector
  data1 <- cbind(meth_control,meth_test)
  num_cluster <- 20
  cluster <-factor(sample (1:num_cluster,dim(data1)[2],replace = TRUE))

  # cluster with similarity=FALSE
  res1 <- SigRM_cluster_test(meth_control,meth_test,unmeth_control,unmeth_test,
  cluster=cluster)

  # cluster with similarity=TRUE
  res2 <- SigRM_cluster_test(meth_control,meth_test,unmeth_control,unmeth_test,
  cluster=cluster,similarity=TRUE)

  # cluster with expression
  expression <- matrix (runif(31000,0,1),nrow=1000,ncol=31)
  res3 <- SigRM_cluster_test(meth_control,meth_test,unmeth_control,unmeth_test,
  cluster=cluster,expression=expression,similarity=TRUE)




  data <- simulateData (test_num=2,control_num=30)

  meth_control=data[[1]];meth_test=data[[2]];unmeth_control=data[[3]];unmeth_test=data[[4]]

  # simulate the cluster vector
  data1 <- cbind(meth_control,meth_test)
  num_cluster <- 20
  cluster <-factor(sample (1:num_cluster,dim(data1)[2],replace = TRUE))

  # cluster with similarity=FALSE
  res1 <- SigRM_cluster_test(meth_control,meth_test,unmeth_control,unmeth_test,
  cluster=cluster)

  # cluster with similarity=TRUE
  res2 <- SigRM_cluster_test(meth_control,meth_test,unmeth_control,unmeth_test,
  cluster=cluster,similarity=TRUE)

  # cluster with expression
  expression <- matrix (runif(32000,0,1),nrow=1000,ncol=32)
  res3 <- SigRM_cluster_test(meth_control,meth_test,unmeth_control,unmeth_test,
  cluster=cluster,
  expression=expression,similarity=TRUE)
}
