% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigRM_similarity_test.R
\name{SigRM_similarity_test}
\alias{SigRM_similarity_test}
\title{This is the function work for the SigRM_cluster_test}
\usage{
SigRM_similarity_test(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  expression = NA,
  method = "Pearson",
  num_control,
  size.factor = NA,
  plot.dispersion = FALSE,
  output.dir = NA,
  remove.false = TRUE,
  q = NA,
  method_dispersion = "locfit",
  adjust_1 = 1e-05,
  adjust_2 = 0.01,
  one_side = TRUE
)
}
\arguments{
\item{meth_control}{detect methylation read counts from false treatment
(row is gene and column is single cell)}

\item{meth_test}{detect methyation read counts from correct treatment
(row is gene and column is single cell)}

\item{unmeth_control}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{unmeth_test}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{expression}{the expression matrix for compare the similarity, if this
variable is miss, we use read counts}

\item{method}{calculate similarity method c("Euclidean distance","Manhattan distance",
"Maximum distance","Pearson","Spearman")}

\item{num_control}{number of cells in the control group}

\item{size.factor}{TRUE or FALSE}

\item{plot.dispersion}{whether plot or not}

\item{output.dir}{the path}

\item{remove.false}{remove some false positive or not}

\item{q}{expression aboundance}

\item{method_dispersion}{the method to estimate the dispersion ("unbiased_estimate","mle","locfit")}

\item{adjust_1}{adjust parameter for RR}

\item{adjust_2}{adjust parameter for OR}

\item{one_side}{p value is one side or both sides}
}
\value{
a list c("methylation level treated","methylation level control","log2 Risk Ratio",
"log2 Odds Ratio","p value","aboundance","adjusted p value","TCR")

1.The first part of the list is the meth proportion data frame (row is gene
and column is single cell) in the test cells.
2.The second part of the list is the mean meth proportion data frame (row is gene
and column is single cell) in the control cells.
3.The third part of the list is the log2 risk ratio data frame for the test cells.
4.The forth part of the list is the log2 odds ratio data frame for the test cells.
5.The fifth part of the list is the p value data frame for the test cells.
6.The sixth part of the list is the estimated gene abundance data frame.
7.The seventh part of the list is the adjusted p value data frame for the test cells.
8.The eighth part of the list is the TCR for the test cells.

For NA in the return list:
\enumerate{
\item NA in the treated methylation level, it is caused by that the test cells have no reads
in both methylation reads and un methylation reads
\item NA in the control methylation level, it is caused by that the all control cells have
no reads in both methylation reads and un methylation reads
\item NA in the risk ratio, the methylation proportion is NA on the test cell or control cells group.
\item NA in the TCR, is similar to the risk ratio.
\item NA in the p value, it is only caused by the that control methylation level is NA and the
treated methylation level is not NA.
}
}
\description{
This function serves as the central component for single-cell RNA methylation
site analysis by selecting the control group based on the similarity of each
test cell. Users can choose the number of control cells, which are selected
based on their similarity to the test cells.
}
\examples{
  data <- simulateData (test_num=1,control_num=30)

  meth_control=data[[1]];meth_test=data[[2]];unmeth_control=data[[3]];unmeth_test=data[[4]]

  # use Pearson
  res1 <- SigRM_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Pearson",num_control=3,size.factor=NA,plot.dispersion=FALSE,
  output.dir = NA,remove.false = TRUE)

  # use Euclidean distance
  res2 <- SigRM_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Euclidean distance",num_control=3,size.factor=NA,
  plot.dispersion=FALSE,   output.dir = NA,remove.false = TRUE)

 # use expression and Euclidean distance
 expression <- matrix (runif(31000,0,1),nrow=1000,ncol=31)
 res3<- SigRM_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
 expression=expression,method="Euclidean distance",num_control=3,
 size.factor=NA,plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)



  data <- simulateData (test_num=7,control_num=30)

  meth_control=data[[1]];meth_test=data[[2]];unmeth_control=data[[3]];unmeth_test=data[[4]]

  # use Pearson
  res1 <- SigRM_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Pearson",num_control=3,size.factor=NA,plot.dispersion=FALSE,
  output.dir = NA,remove.false = TRUE)

  # use Euclidean distance
  res2 <- SigRM_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
  method="Euclidean distance",num_control=3,size.factor=NA,
  plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)



 # use expression and Euclidean distance
 expression <- matrix (runif(37000,0,1),nrow=1000,ncol=37)
 res5<- SigRM_similarity_test(meth_control,meth_test,unmeth_control,unmeth_test,
 expression=expression,method="Euclidean distance",num_control=3,
 size.factor=NA,plot.dispersion=FALSE, output.dir = NA,remove.false = TRUE)

}
