% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data function}
\usage{
simulateData(
  n_Sites = 2000,
  test_num = 100,
  control_num = 300,
  min_expression = 0,
  max_expression = 4,
  expression_p = 10,
  per_me = 1/2,
  prange1 = c(0, 0.4),
  prange2 = c(0.6, 0.8),
  prange3 = c(0, 0.4),
  shape2 = 10
)
}
\arguments{
\item{n_Sites}{number of site}

\item{test_num}{number of sample in correct treatment}

\item{control_num}{number of sample in false treatment}

\item{min_expression}{min expression for abundance}

\item{max_expression}{max expression for abundance}

\item{expression_p}{expression parameter}

\item{per_me}{percentage of difference}

\item{prange1}{parameter for simulate p1}

\item{prange2}{parameter for simulate p2}

\item{prange3}{parameter for simulate p3}

\item{shape2}{the hyper-parameter of the beta distribution}
}
\value{
simulate read counts( meth read counts of control data,
meth read counts of test data, unmeth read counts of control data,
unmeth read counts of test data)
}
\description{
Simulate data function
}
\examples{
 data <- simulateData (test_num=10,control_num=30)


}
