% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sizeFactor}
\alias{sizeFactor}
\title{This is the document that contains the sub function for the main function
Calculate the size factors.}
\usage{
sizeFactor(data)
}
\arguments{
\item{data}{read counts}
}
\value{
size factor
}
\description{
This is the document that contains the sub function for the main function
Calculate the size factors.
}
