% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGMRtest.R
\name{SIGMRtest}
\alias{SIGMRtest}
\title{This function serves as the central component for single-cell RNA methylation
site analysis by clustering cells based on gene expression or read counts,
encompassing both test and control groups.}
\usage{
SIGMRtest(
  meth_control,
  meth_test,
  unmeth_control,
  unmeth_test,
  size.factor = NA,
  plot.dispersion = FALSE,
  output.dir = NA,
  remove.false = TRUE
)
}
\arguments{
\item{meth_control}{detect methylation read counts from false treatment
(row is gene and column is single cell)}

\item{meth_test}{detect methyation read counts from correct treatment
(row is gene and column is single cell)}

\item{unmeth_control}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{unmeth_test}{detect unmethylation read counts from false treatment
(row is gene and column is single cell)}

\item{size.factor}{NA or a data.frame with one column is control size factor and test size factor}

\item{plot.dispersion}{whether plot or not}

\item{output.dir}{the path}

\item{remove.false}{remove some false positive or not}
}
\value{
a list c("detect proportion treated","detect proportion control","log2 Risk Ratio",
"log2 Odds Ratio","p value","abundance","adjusted p value")

1.The first part of the list is the meth proportion data frame (row is gene
and column is single cell) in the test cells.
2.The second part of the list is the mean meth proportion vector in the control cells.
3.The third part of the list is the log2 risk ratio data frame for the test cells.
4.The forth part of the list is the log2 odds ratio data frame for the test cells.
5.The fifth part of the list is the p value data frame for the test cells.
6.The sixth part of the list is the estimated gene abundance vector.
7.The seventh part of the list is the adjusted p value data frame for the test cells.
}
\description{
This function serves as the central component for single-cell RNA methylation
site analysis by clustering cells based on gene expression or read counts,
encompassing both test and control groups.
}
\examples{
  data <- simulateData (test_num=1,control_num=30)


  res <- SIGMRtest(data[[1]],data[[2]],data[[3]],data[[4]])

  data <- simulateData (test_num=2,control_num=30)


  res <- SIGMRtest(data[[1]],data[[2]],data[[3]],data[[4]])

}
