% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data function}
\usage{
simulateData(
  n_Sites = 2000,
  test_num = 100,
  control_num = 300,
  min_expression = 1,
  max_expression = 4,
  per_me = 1/2,
  prange1 = c(0, 0.2),
  p_s_1 = 0.01,
  prange2 = c(0.25, 0.3),
  prange3 = c(0, 0.3),
  p_s_3 = 0.01
)
}
\arguments{
\item{n_Sites}{number of site}

\item{test_num}{number of sample in correct treatment}

\item{control_num}{number of sample in false treatment}

\item{min_expression}{min expression for abundance}

\item{max_expression}{max expression for abundance}

\item{per_me}{percentage of difference}

\item{prange1}{parameter for simulate p1}

\item{p_s_1}{parameter for step of p1}

\item{prange2}{parameter for simulate p2}

\item{prange3}{parameter for simulate p3}

\item{p_s_3}{parameter for step of p3}
}
\value{
simulate read counts( meth read counts of control data,
meth read counts of test data, unmeth read counts of control data,
unmeth read counts of test data)
}
\description{
Simulate data function
}
\examples{
 data <- simulateData (test_num=10,control_num=30)


}
